refactor: update data type and modify file name
